fixed bug and implemented features

+++ internal/domain/models/refreshToken.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type RefreshToken struct {
-	Id         int
-	UserId     int
+	Id         int64
+	UserId     int64
 	Token      string
 	CreateDate time.Time
 	ExpDate    time.Time
